[WIP] Implement cross-reference to eaxml
This commit is for implementing cross-reference from eatxt to eaxml,
the idea is borrowed from the code in ..ui.handlers. The feature itself
has been implemented, however there is an issue when debugging, so it
is should be a [WIP] commit. Here I added two methods for this imple-
mentation which are turned off when being merged in, so it won't affect
current feature.