Added custom scope to accommodate different URI schemes
When serialising an EMF model that was loaded from an EAXML file to an
EAtxt file, Xtext uses "file" as the scheme for the EObject URIs. The
reason is that Xtext is not aware of EastADLURIFactory and there is no
simple way (that I could find) to make it aware of that. Instead, we are
now using a specialised implementation of Scope, that compares the URI
fragments if the URI of the EObject that we look for in the scope has
"ea" as its scheme. This allows us to find the correct element since
East-ADL uses a global namespace (confirm!).