Updated target platform to use Eclipse 2021-03
This has several consequences:
 * We now use Xtext 2.25.0 which was the last version that was part of
Simrel and the current version at time of writing; this is also the main
reason for the update.
 * We are now using an Eclipse version that supports (and requires) Java
11.
 * When integrating with EATOP, we need to use the version based on the
Eclipse 2021-03 platform; there's a pull request for that on Bitbucket.

Note: we are still using Apache Xerces from Orbit 2019-12 since the
version in Orbit 2021-03 is not compatible with Eclipse Sphinx.