Refactored scope provider to use the resource set to access all models
Previously, we were parsing eatxt files directly. We are now using the
resource set to access these files as well, making the API more simple
and using the same standard approach for both EAXML and EATXT files.
In addition, we are now compiling the scope from all resources in the
workspace, including ones that live in subfolders. We also only need to
traverse all files once and compare to a list of file extensions to find
relevant files.