Fix #26: UUIDs are now treated as strings surrounded by quotes
The UUIDs used in EAST-ADL are of the form
	('a'..'f'|'A'..'F'|'0'..'9') ('a'..'f'|'A'..'F'|'0'..'9'|'-')*
Unfortunately, we cannot use this rule as a terminal since it conflicts
with other rules and makes the grammar ambiguous. We are therefore now
using a different approach where the UUIDs are treated as strings that
are surrounded by quotes. To make this work, an additional
ValueConverter is necessary to avoid that Xtext tries to treat the rule
as an ID (which it does when using the built in rules for Strings).