Scope now ignores query when looking for elements
The custom scope we are using is already aware of EAST-ADL URIs. When
serialising, however, the URIs also contain a query element. These
queries differ under certain circumstances which means that the elements
are not found. This commit trims the queries from the URIs, allowing the
scope to find the needed elements.