Fix #33: Scope Provider now Looks Up Arbitrary Cross References
Added an Activator that creates one single time on startup a lookup map
containing all cross references from the EAST-ADL metamodel. The
EaTextScopeProvider now gets this lookup map on every auto-completion
keystroke and searches it, instead of conducting a multitude of name
comparisons for context and target class pairs. These were bad from a
maintenance as well as a performance point of view.