MWE2 now uses metamodel correctly
The MWE2 workflow had a hack that used the genmodel from within a jar
file for generation. This was always just a placeholder since it made
evolution of the metamodel unwieldy. This commit adds a less hacky
solution where we explicitly register the project in which the
meta-model resides so that Xtext picks it up without problem.