Fix Generated Parser w/ too many Constants Issue due to Larger Grammar
After we went to the full EAST-ADL metamodel and thereby to a larger
grammar, the default parser generation yielded a parser class that did
not compile due to the fact that it contained too many constants. With
this different parser generator option (classSplitting = true), the
parser is generated in a different way so that it compiles again.