Forcibly added src-gen folder
When doing a clean checkout, Eclipse does not set the classpath
correctly since the src-gen folder in the org.bumble.eatxt project is
missing. That means that we cannot run the MWE2 workflow to generate the
Xtext artifacts.

This commit forcibly adds this folder which makes it easier for new
users to get started since they do not have to worry about setting the
classpath before generating artifacts.