Fix #32: Content assist for short names now automatically generated
Previously, we were using a manually generated class to provide content
assist for auto-completion of short names for model elements. This
commit adds a generator fragment that creates the methods for these
proposals automatically. Using the new generator also made it necessary
to further customize the MWE2 workflow which is now based on a custom
language rather than the standard language that removes the original
content assist fragment.