Removed "new file" Figure associated with Old Name and added Eatext one
The old content type was called "EastAdlSimplified", so the new file
figure was associated with this name. After we renamed it to "Eatxt",
the name of the new file figure also changed. This fix removes the
obsolete figure and adds the same figure with the new name, so that VCS
changes are not indicated anymore after each Xtext infrastructure
generation.