First Shot on Templates for Issue #34
Created an exemplary template for the production rule ModeGroup, which
has a mandatory attribute precondition and at least one mandatory nested
element Mode, which in turn has another mandatory attribute condition.
No templates for the full language yet, because I think this approach is
very interesting but a lot has to be discussed before spending the
effort on a fully fledged approach.