Support String-based Attributes in Generated Templates (#34)
Attributes with types EString or STRING (which we now have quite a lot
of) were not correctly proposed by the generated templates, because the
corresponding quotation marks were missing. Now the generator checks
whether the type in lower case contains a "string", and surrounds in the
positive case its value with quotation marks.