Changed Two Attributes Optional to Improve EAXML->EAXML Conversion (#36)
The EAXML->EAXML conversion for the two metamodel subset examples
"BasicWW...eaxml" failed, because the attributes
RangeableValueType#baseRangeable and FunctionFlowPort#type were not
optional in the grammar. This is due to the fact that they are mandatory
in the metamodel. However, EATOP relaxes these mandatory attributes so
that it is possible to create EAXML files without these attributes,
which in turn failed the conversion. See also comments in the discussion
here.

However, the conversion does not yet work for the large example
FMU_BBW_4Wheel_2.2EM.eaxml using the full metamodel.